🩹 Remove broken `forward` implementation
This library has to implement its own version of `std::forward` so that
the `not_null` implementation may be `constexpr` even when compiling in
C++11 (which did not have this defined as `constexpr`). This
implementation was unfortunately broken in C++14 onward, due to the
`__cplusplus` branch using the identifier `forward` rather than
`not_null_forward`.

This fixes the problem by removing the call to `std::forward` entirely
and just committing to our own implementation instead.