✨ Suppress faulty `-Wunused-variable`
The `check_not_null` function makes use of comma-operator sequencing,
which is triggering "unused variable" warnings due to the first
expression being "ignored" as it proceeds to the return value.

This behavior is done intentionally so that the non-constexpr
sub-expression that throws does not get evaluated (as it shouldn't),
and if ever it *does* get triggered, it will result in an exception.

Thus, the result isn't actually "unused"; just the compiler diagnoses
that it is.

This uses compiler pragmas to suppress this behavior.