🔨 Refactor `not_null` factory pattern
`not_null` is currently authored to use `friend`ship so that the
`assume_not_null` may invoke a private/hidden constructor within
`not_null`. This pattern works on most compilers, however the complex
template expression for the friendship is resuling in an internal
compiler error occurring on some versions of GCC.

To simplify the friendship expression, an aggregate type has been
created that solely exists to behave as a `friend` to `not_null`:
the `not_null_factory`. This allows a complex `friend` declaration
to turn into a `friend` statement that names a type -- which results in
less errors.