🩹 Fix incorrect C++14 `constexpr`
The implementation function for throwing a null contract violation
error was marked `constexpr` in C++14, but was mistakenly not
implemented as `constexpr` -- resulting in errors when building above
C++11.

Since this can never produce a proper core constant expression (due to
throwing an exception), this function  should never be marked `constexpr`
to begin with; so this has been removed.