🔨 Suppress `clang-cl`'s `-Wcomma` diagnostic
`clang-cl` is warning about a potential misuse of the comma operator,
although this diagnostic is invalid. The comma operator is being
used appropriately as a sequencing method, which is guaranteed by the
C++ standard for the builtin comma operator effect.

To suppress this issue, a compiler `pragma` has been used.