🩹 Guard `-Wcomma` suppression for `clang`
A previous commit introduced a suppression for the "-Wcomma" diagnostic,
which is incorrectly emitted in some versions of clang when enabled.

However, it appears that `-Wcomma` was only introduced in 3.9 onward;
and thus otherwise trigger an error on earlier versions of `clang`. To
avoid potential compilation failures, this suppression has now been
guarded by version checks.