🔨 Address MSVC failure
MSVC, as usual, is refusing to accept completely valid C++ code --
despite compiling on the most recent MSVC version.

The `not_null_to_address` function is being incorrectly deduced as
not having a compatible definition, as according to MSVC -- despite the
fact that the definition is clearly visible. Moving the out-of-line
definition into an inline definition appears to address this diagnostic
for them.

The code has been updated to spoon-feed this to MSVC.