🩹 Guard exception definition when disabled
When the preprocessor token `NOT_NULL_EXCEPTIONS_DISABLED` is defined,
all exception-related code -- including the exception itself -- should
no longer be emitted. However, a bug in the implementation is
resulting in the _definition_ of the constructor to be present, while
the class declaration is not present -- resulting in compile-errors.

This guards the constructor body definition as well so that this will
not occur.