🔨 Hint to compiler that `not_null` is not null
The `not_null` object prevents the possibility of the underlying pointer
being null. This is information that the C++ compiler can often leverage
for optimizations if it can determine such cases.

The code has been documented with C++ attributes for GCC and clang, and
using `__assume` for MSVC, to attempt to hint to the compiler that the
pointer contained in a `not_null` can never legally be null.